% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_seg.R
\name{canopy_seg}
\alias{canopy_seg}
\title{Calculate visual obstruction of specific vertical canopy sections}
\usage{
canopy_seg(
  input_path,
  image_output = FALSE,
  low_top = 0.22,
  low_bot = 0,
  mid_top = 0.43,
  mid_bot = 0.22,
  upper_top = 1,
  upper_bot = 0.43,
  low_output = NULL,
  mid_output = NULL,
  upper_output = NULL
)
}
\arguments{
\item{input_path}{File path for binary input images}

\item{image_output}{Logical operator to enable visual output, Default = FALSE}

\item{low_top}{Numeric value of upper limit of lowest canopy section, Default = 0.22}

\item{low_bot}{Numeric value of lower limit of lowest canopy section, Default = 0.00}

\item{mid_top}{Numeric value of upper limit of middle canopy section, Default = 0.43}

\item{mid_bot}{Numeric value of lower limit of middle canopy section, Default = 0.22}

\item{upper_top}{Numeric value of upper limit of upper canopy section, Default = 1.00}

\item{upper_bot}{Numeric value of lower limit of upper canopy section, Default = 0.43}

\item{low_output}{File path for lower canopy binary output images when image_output = TRUE, Default = NULL}

\item{mid_output}{File path for middle canopy binary output images when image_output = TRUE, Default = NULL}

\item{upper_output}{File path for upper canopy binary output images when image_output = TRUE, Default = NULL}
}
\value{
A dataframe with estimated vegetation visual obstruction for specified canopy sections and binary images if selected
}
\description{
Calculate visual obstruction of specific vertical canopy sections
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/vo_img",sep ="", collapse="")
sample_canopy_df <- canopy_seg(input_path = sample_imgs)
}
