% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_seg.R
\name{fuel_seg}
\alias{fuel_seg}
\title{Segment and calculate percent visual obstruction of live and dead fuel}
\usage{
fuel_seg(
  input_path,
  image_output = FALSE,
  output_path = NULL,
  veg_vo = NULL,
  training,
  thresh,
  ntree,
  mtry
)
}
\arguments{
\item{input_path}{File path for input images}

\item{image_output}{Logical operator to enable visual output, Default = FALSE}

\item{output_path}{File path for binary output images when image_output = TRUE, Default = NULL}

\item{veg_vo}{List of values for associated vegetation visual obstruction values to calculate proportion live fuel}

\item{training}{Training data file for Random Forest Model}

\item{thresh}{Numeric value of threshold value for Random Forest prediction}

\item{ntree}{Number of trees to grow for Random Forest Model}

\item{mtry}{Number of variables to randomly sample as candidates at each split}
}
\value{
a data frame with estimated live fuel visual obstruction (percentage of overall vegetation) and binary images if selected
}
\description{
Segment and calculate percent visual obstruction of live and dead fuel
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/img",sep ="", collapse="")
training_file <- system.file("stillwater.training.rda")
sample_lf_vo_df <- lf_seg(input_path = sample_imgs,
                            image_output = FALSE,
                            training = training_file,
                            thresh = 0.80,
                            ntree = 100,
                            mtry = 1)
}
