% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm_test.R
\name{rfm_test}
\alias{rfm_test}
\title{Returns error and convergence estimates for Random Forest Model based on training data}
\usage{
rfm_test(training, focal_column, thresh, ntree, mtry)
}
\arguments{
\item{training}{Training data file for Random Forest Model}

\item{focal_column}{Column within training data file denoting focal vegetation}

\item{thresh}{Numeric value of threshold value for Random Forest prediction}

\item{ntree}{Number of trees to grow for Random Forest Model}

\item{mtry}{Number of variables to randomly sample as candidates at each split}
}
\value{
A print summary of estimated model error and plot of error as a function number of trees
}
\description{
Returns error and convergence estimates for Random Forest Model based on training data
}
