% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_ht.R
\name{veg_ht}
\alias{veg_ht}
\title{Calculate vegetation height from segmented images}
\usage{
veg_ht(input_path, board_height, veg_type)
}
\arguments{
\item{input_path}{File path for binary input images}

\item{board_height}{Numeric value of actual height of profile board}
}
\value{
A dataframe with mean, maximum, and standard deviation of vegetation height for input images
}
\description{
Calculate vegetation height from segmented images
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/vo_img",sep ="", collapse="")
sample_heights_df <- veg_ht(input_path = ,
                            board_height = 116)
}
