% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_seg.R
\name{veg_seg}
\alias{veg_seg}
\title{Segment vegetation from background and calculate visual obstruction}
\usage{
veg_seg(
  input_path,
  image_output = FALSE,
  output_path = NULL,
  training,
  thresh,
  ntree,
  mtry
)
}
\arguments{
\item{input_path}{File path for input images}

\item{image_output}{= Logical operator to enable visual output, Default = FALSE}

\item{output_path}{File path for binary output images when image_output = TRUE, Default = NULL}

\item{training}{Training data file for Random Forest Model}

\item{thresh}{Numeric value of threshold value for Random Forest prediction}

\item{ntree}{Number of trees to grow for Random Forest Model}

\item{mtry}{Number of variables to randomly sample as candidates at each split}
}
\value{
A dataframe with estimated vegetation visual obstruction and binary images if selected
}
\description{
Segment vegetation from background and calculate visual obstruction
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/img",sep ="", collapse="")
training_file <- system.file("stillwater.training.rda")
sample_veg_vo_df <- veg_seg(input_path = sample_imgs,
                            image_output = FALSE,
                            training = training_file,
                            thresh = 0.95,
                            ntree = 100,
                            mtry = 1)
}
