% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_var.R
\name{veg_var}
\alias{veg_var}
\title{Calculate horizontal variance of vegetation visual obstruction within profile board images}
\usage{
veg_var(
  input_path,
  image_output = FALSE,
  left_output = NULL,
  center_output = NULL,
  right_output = NULL
)
}
\arguments{
\item{input_path}{File path for binary input images}

\item{image_output}{Logical operator to enable visual output, Default = FALSE}

\item{left_output}{File path for left section binary output images when image_output = TRUE, Default = NULL}

\item{center_output}{File path for center section binary output images when image_output = TRUE, Default = NULL}

\item{right_output}{File path for right section binary output images when image_output = TRUE, Default = NULL}
}
\value{
A dataframe with estimated vegetation visual obstruction for specified horizontal sections, variance, and standard deviation and binary images if selected
}
\description{
Calculate horizontal variance of vegetation visual obstruction within profile board images
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/vo_img",sep ="", collapse="")
sample_variance_df <- veg_var(sample_imgs)
}
