% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_seg.R
\name{fuel_seg}
\alias{fuel_seg}
\title{Segment and calculate percent visual obstruction of live and dead fuel}
\usage{
fuel_seg(
  input_path,
  lf_output = FALSE,
  lf_path = NULL,
  df_output = FALSE,
  df_path = NULL,
  veg_vo = NULL,
  training,
  veg_thresh,
  fuel_thresh,
  ntree,
  mtry
)
}
\arguments{
\item{input_path}{File path for input images}

\item{lf_output}{Logical operator to enable live fuel visual output, Default = FALSE}

\item{lf_path}{File path for binary output images for live fuel when lf_output = TRUE, Default = NULL}

\item{df_output}{Logical operator to enable dead fuel visual output, Default = FALSE}

\item{df_path}{File path for binary output images for dead fuel when df_output = TRUE, Default = NULL}

\item{veg_vo}{Option to provide overall vegetation obstruction values (list) otherwise function will calculate, Default = NULL}

\item{training}{Training data file for Random Forest Model}

\item{veg_thresh}{Numeric value of threshold value for overall vegetation Random Forest prediction}

\item{fuel_thresh}{Numeric value of threshold value for fuel type Random Forest prediction}

\item{ntree}{Number of trees to grow for Random Forest Model}

\item{mtry}{Number of variables to randomly sample as candidates at each split}
}
\value{
a data frame with estimated live fuel visual obstruction (percentage of overall vegetation) and binary images if selected
}
\description{
Segment and calculate percent visual obstruction of live and dead fuel
}
\examples{
sample_imgs <- paste(find.package(package="heterogeneityR"),"/inst/img",sep ="", collapse="")
training_file <- system.file("stillwater.training.rda")
sample_lf_vo_df <- lf_seg(input_path = sample_imgs,
                            image_output = FALSE,
                            training = training_file,
                            thresh = 0.80,
                            ntree = 100,
                            mtry = 1)
}
